:art: Cosmetic changes and updating attribute type from int to float.

+++ src/entities/employee.py
@@ -15,15 +15,13 @@ class Employee:
 
 	name: str
 	employee_id: int
-	pay_rate: float = 100.00
-	hours_worked: int = 0
-	employer_cost: float = 1000.00
+	pay_rate: float = 100.0
+	hours_worked: float = 0.0
+	employer_cost: float = 1000.0
 	has_comission: bool = True
-	comission: float = 100.00
+	comission: float = 100.0
 	contracts_landed: int = 0
 
 	def compute_payout(self) -> float:
 		"""Compute how much must be paid to an employee"""
 		raise NotImplementedError()
-	
-
